requirements.txt
agate==1.9.1
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
daff==1.3.46
dbt-adapters==1.5.0
dbt-common==1.8.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
dbt-snowflake==1.8.3
deepdiff==7.0.1
filelock==3.16.0
h11==0.14.0
idna==3.8
importlib-metadata==6.11.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.3.0
leather==0.4.0
Logbook==1.5.3
MarkupSafe==2.1.5
mashumaro==3.13.1
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.3
numpy==2.1.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
panda==0.3.1
pandas==2.2.2
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.3.2
protobuf==4.25.4
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
selenium==4.24.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tomlkit==0.13.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.20.1
astronomer-cosmos
apache-airflow-providers-snowflake